move envoy package to test/, avoid var shadowing
