+++ backend/module/chaos/serverexperimentation/rtds/integration/config/generate_config.go
@@ -5,7 +5,7 @@ import (
 
 	_ "github.com/envoyproxy/go-control-plane/envoy/extensions/filters/http/fault/v3"
 
-	"github.com/lyft/clutch/backend/module/chaos/serverexperimentation/rtds/envoy"
+	"github.com/lyft/clutch/backend/test/envoy"
 )
 
 const faultFilterConfig = `
@@ -52,7 +52,7 @@ type: STRICT_DNS
 `
 
 func main() {
-	config := envoy.NewEnvoyConfig()
+	config := testenvoy.NewEnvoyConfig()
 
 	err := config.AddCluster(rtdsCluster)
 	if err != nil {
